finopti-platform/sub_agents/code_execution_agent_adk/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
google-adk>=0.1.0
google-genai>=0.2.0
google-cloud-secret-manager>=2.16.0
google-cloud-bigquery>=3.10.0
arize-phoenix>=4.0.0
openinference-instrumentation-google-adk>=0.1.0
opentelemetry-sdk
opentelemetry-exporter-otlp

redis>=5.0.0
pydantic>=2.0.0
